Refactor: Rename FieldType namespaces to Field across schema and generator classes
