+++ src/Service/Lexicon/V1/ComponentGenerator/Field/ArrayComponentGenerator.php
@@ -4,7 +4,7 @@ namespace Blugen\Service\Lexicon\V1\ComponentGenerator\Field;
 
 use Blugen\Service\Lexicon\GeneratorInterface;
 use Blugen\Service\Lexicon\V1\Property;
-use Blugen\Service\Lexicon\V1\TypeSpecificSchema\FieldType\ArraySchema;
+use Blugen\Service\Lexicon\V1\TypeSpecificSchema\Field\ArraySchema;
 use Nette\PhpGenerator\ClassType;
 
 class ArrayComponentGenerator implements GeneratorInterface
