+++ src/Service/Lexicon/V1/ComponentGenerator/Field/BooleanComponentGenerator.php
@@ -4,7 +4,7 @@ namespace Blugen\Service\Lexicon\V1\ComponentGenerator\Field;
 
 use Blugen\Service\Lexicon\GeneratorInterface;
 use Blugen\Service\Lexicon\V1\Property;
-use Blugen\Service\Lexicon\V1\TypeSpecificSchema\FieldType\BooleanSchema;
+use Blugen\Service\Lexicon\V1\TypeSpecificSchema\Field\BooleanSchema;
 use Nette\PhpGenerator\ClassType;
 
 class BooleanComponentGenerator implements GeneratorInterface
