+++ src/Service/Lexicon/V1/ComponentGenerator/Field/BytesComponentGenerator.php
@@ -4,7 +4,7 @@ namespace Blugen\Service\Lexicon\V1\ComponentGenerator\Field;
 
 use Blugen\Service\Lexicon\GeneratorInterface;
 use Blugen\Service\Lexicon\V1\Property;
-use Blugen\Service\Lexicon\V1\TypeSpecificSchema\FieldType\BytesSchema;
+use Blugen\Service\Lexicon\V1\TypeSpecificSchema\Field\BytesSchema;
 use Nette\PhpGenerator\ClassType;
 
 class BytesComponentGenerator implements GeneratorInterface
