+++ src/Service/Lexicon/V1/ComponentGenerator/Field/CidLinkComponentGenerator.php
@@ -4,7 +4,7 @@ namespace Blugen\Service\Lexicon\V1\ComponentGenerator\Field;
 
 use Blugen\Service\Lexicon\GeneratorInterface;
 use Blugen\Service\Lexicon\V1\Property;
-use Blugen\Service\Lexicon\V1\TypeSpecificSchema\FieldType\CidLinkSchema;
+use Blugen\Service\Lexicon\V1\TypeSpecificSchema\Field\CidLinkSchema;
 use Nette\PhpGenerator\ClassType;
 
 class CidLinkComponentGenerator implements GeneratorInterface
