+++ src/Service/Lexicon/V1/ComponentGenerator/Field/IntegerComponentGenerator.php
@@ -4,7 +4,7 @@ namespace Blugen\Service\Lexicon\V1\ComponentGenerator\Field;
 
 use Blugen\Service\Lexicon\GeneratorInterface;
 use Blugen\Service\Lexicon\V1\Property;
-use Blugen\Service\Lexicon\V1\TypeSpecificSchema\FieldType\IntegerSchema;
+use Blugen\Service\Lexicon\V1\TypeSpecificSchema\Field\IntegerSchema;
 use Nette\PhpGenerator\ClassType;
 
 class IntegerComponentGenerator implements GeneratorInterface
