+++ src/Service/Lexicon/V1/ComponentGenerator/Field/TokenComponentGenerator.php
@@ -4,7 +4,7 @@ namespace Blugen\Service\Lexicon\V1\ComponentGenerator\Field;
 
 use Blugen\Service\Lexicon\GeneratorInterface;
 use Blugen\Service\Lexicon\V1\Property;
-use Blugen\Service\Lexicon\V1\TypeSpecificSchema\FieldType\TokenSchema;
+use Blugen\Service\Lexicon\V1\TypeSpecificSchema\Field\TokenSchema;
 use Nette\PhpGenerator\ClassType;
 
 class TokenComponentGenerator implements GeneratorInterface
