+++ src/Service/Lexicon/V1/DefGenerator/Field/ArrayGenerator.php
@@ -6,7 +6,7 @@ use Blugen\Service\Lexicon\GeneratorInterface;
 use Blugen\Service\Lexicon\V1\Resolver\NamespaceResolver;
 use Blugen\Service\Lexicon\V1\Schema;
 use Blugen\Service\Lexicon\V1\Support\ArrayHandler;
-use Blugen\Service\Lexicon\V1\TypeSpecificDefinition\FieldType\ArrayTypeDefinition;
+use Blugen\Service\Lexicon\V1\TypeSpecificDefinition\Field\ArrayTypeDefinition;
 use Nette\PhpGenerator\ClassType;
 use Nette\PhpGenerator\PhpFile;
 use Nette\PhpGenerator\PhpNamespace;
