+++ src/Service/Lexicon/V1/TypeSpecificSchema/Field/BlobSchema.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Blugen\Service\Lexicon\V1\TypeSpecificSchema\FieldType;
+namespace Blugen\Service\Lexicon\V1\TypeSpecificSchema\Field;
 
 use Blugen\Service\Lexicon\SchemaInterface;
 
-class NullSchema implements SchemaInterface
+class BlobSchema implements SchemaInterface
 {
     public function __construct(
         private readonly SchemaInterface $schema
@@ -12,12 +12,22 @@ class NullSchema implements SchemaInterface
 
     public function type(): string
     {
-        return 'null';
+        return 'blob';
     }
 
     public function description(): ?string
     {
         return $this->schema->description();
+    }
+
+    public function accept(): ?array
+    {
+        return $this->__get('accept') ?? null;
+    }
+
+    public function maxSize(): ?int
+    {
+        return $this->__get('maxSize') ?? null;
     }
 
     public function __get(string $name): mixed
