+++ src/Service/Lexicon/V1/TypeSpecificSchema/Field/BooleanSchema.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Blugen\Service\Lexicon\V1\TypeSpecificSchema\FieldType;
+namespace Blugen\Service\Lexicon\V1\TypeSpecificSchema\Field;
 
 use Blugen\Service\Lexicon\SchemaInterface;
 
-class IntegerSchema implements SchemaInterface
+class BooleanSchema implements SchemaInterface
 {
     public function __construct(private readonly SchemaInterface $schema)
     {}
@@ -24,27 +24,12 @@ class IntegerSchema implements SchemaInterface
         return $this->schema->description() ?? null;
     }
 
-    public function minimum(): ?int
-    {
-        return $this->__get('minimum');
-    }
-
-    public function maximum(): ?int
-    {
-        return $this->__get('maximum');
-    }
-
-    public function enum(): ?array
-    {
-        return $this->__get('enum');
-    }
-
-    public function default(): ?int
+    public function default(): ?bool
     {
         return $this->__get('default');
     }
 
-    public function const(): ?int
+    public function const(): ?bool
     {
         return $this->__get('const');
     }
