+++ src/Service/Lexicon/V1/TypeSpecificSchema/Field/NullSchema.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Blugen\Service\Lexicon\V1\TypeSpecificSchema\FieldType;
+namespace Blugen\Service\Lexicon\V1\TypeSpecificSchema\Field;
 
 use Blugen\Service\Lexicon\SchemaInterface;
 
-class BlobSchema implements SchemaInterface
+class NullSchema implements SchemaInterface
 {
     public function __construct(
         private readonly SchemaInterface $schema
@@ -12,22 +12,12 @@ class BlobSchema implements SchemaInterface
 
     public function type(): string
     {
-        return 'blob';
+        return 'null';
     }
 
     public function description(): ?string
     {
         return $this->schema->description();
-    }
-
-    public function accept(): ?array
-    {
-        return $this->__get('accept') ?? null;
-    }
-
-    public function maxSize(): ?int
-    {
-        return $this->__get('maxSize') ?? null;
     }
 
     public function __get(string $name): mixed
