+++ src/Service/Lexicon/V1/TypeSpecificSchema/Field/ObjectSchema.php
@@ -1,16 +1,14 @@
 <?php
 
-namespace Blugen\Service\Lexicon\V1\TypeSpecificSchema\FieldType;
+namespace Blugen\Service\Lexicon\V1\TypeSpecificSchema\Field;
 
 use Blugen\Service\Lexicon\SchemaInterface;
 use Blugen\Service\Lexicon\V1\Property;
 use Blugen\Service\Lexicon\V1\Schema;
 
-class ParamsSchema implements SchemaInterface
+class ObjectSchema implements SchemaInterface
 {
-    public function __construct(
-        private readonly SchemaInterface $schema,
-    )
+    public function __construct(private readonly SchemaInterface $schema)
     {
     }
 
@@ -21,7 +19,7 @@ class ParamsSchema implements SchemaInterface
 
     public function description(): ?string
     {
-        return $this->schema->description();
+        return $this->schema->description() ?? null;
     }
 
     public function __get(string $name): mixed
@@ -29,27 +27,27 @@ class ParamsSchema implements SchemaInterface
         return $this->schema->__get($name);
     }
 
-    /**
-     * @return string[]|null
-     */
-    public function required(): ?array
-    {
-        return $this->__get('required');
-    }
-
-    /**
-     * @return Property[]
-     */
     public function properties(): array
     {
+        $nullable = $this->nullable() ?? [];
         $required = $this->required() ?? [];
         $properties = $this->__get('properties');
 
         return array_map(fn (string $name, array $rawSchema) => new Property(
             $name,
             new Schema($rawSchema),
-            ! in_array($name, $required, true),
+            in_array($name, $nullable, true),
             in_array($name, $required, true),
         ), array_keys($properties), $properties);
+    }
+
+    public function required(): ?array
+    {
+        return $this->__get('required');
+    }
+
+    public function nullable(): ?array
+    {
+        return $this->__get('nullable');
     }
 }
