+++ src/Service/Lexicon/V1/TypeSpecificSchema/Field/ParamsSchema.php
@@ -1,14 +1,16 @@
 <?php
 
-namespace Blugen\Service\Lexicon\V1\TypeSpecificSchema\FieldType;
+namespace Blugen\Service\Lexicon\V1\TypeSpecificSchema\Field;
 
 use Blugen\Service\Lexicon\SchemaInterface;
 use Blugen\Service\Lexicon\V1\Property;
 use Blugen\Service\Lexicon\V1\Schema;
 
-class ObjectSchema implements SchemaInterface
+class ParamsSchema implements SchemaInterface
 {
-    public function __construct(private readonly SchemaInterface $schema)
+    public function __construct(
+        private readonly SchemaInterface $schema,
+    )
     {
     }
 
@@ -19,7 +21,7 @@ class ObjectSchema implements SchemaInterface
 
     public function description(): ?string
     {
-        return $this->schema->description() ?? null;
+        return $this->schema->description();
     }
 
     public function __get(string $name): mixed
@@ -27,27 +29,27 @@ class ObjectSchema implements SchemaInterface
         return $this->schema->__get($name);
     }
 
+    /**
+     * @return string[]|null
+     */
+    public function required(): ?array
+    {
+        return $this->__get('required');
+    }
+
+    /**
+     * @return Property[]
+     */
     public function properties(): array
     {
-        $nullable = $this->nullable() ?? [];
         $required = $this->required() ?? [];
         $properties = $this->__get('properties');
 
         return array_map(fn (string $name, array $rawSchema) => new Property(
             $name,
             new Schema($rawSchema),
-            in_array($name, $nullable, true),
+            ! in_array($name, $required, true),
             in_array($name, $required, true),
         ), array_keys($properties), $properties);
-    }
-
-    public function required(): ?array
-    {
-        return $this->__get('required');
-    }
-
-    public function nullable(): ?array
-    {
-        return $this->__get('nullable');
     }
 }
